pyproject.toml
[project]
name = "hunyuanimage3"
description = "Professional ComfyUI custom nodes for Tencent's HunyuanImage-3.0 80B multimodal model — text-to-image, Instruct editing, multi-image fusion, block swap, multiple quantization modes (BF16/INT8/NF4), smart memory management, and high-res generation."
version = "1.2.0"
license = {file = "LICENSE"} 
# classifiers = [
#     # For OS-independent nodes (works on all operating systems)
#     "Operating System :: OS Independent",
#
#     # OR for OS-specific nodes, specify the supported systems:
#     "Operating System :: Microsoft :: Windows",  # Windows specific
#     "Operating System :: POSIX :: Linux",  # Linux specific
#     "Operating System :: MacOS",  # macOS specific
#
#     # GPU Accelerator support. Pick the ones that are supported by your extension.
#     "Environment :: GPU :: NVIDIA CUDA",    # NVIDIA CUDA support
#     "Environment :: GPU :: AMD ROCm",       # AMD ROCm support
#     "Environment :: GPU :: Intel Arc",      # Intel Arc support
#     "Environment :: NPU :: Huawei Ascend",  # Huawei Ascend support
#     "Environment :: GPU :: Apple Metal",    # Apple Metal support
# ]

dependencies = ["transformers>=4.47.0", "torch>=2.8.0", "bitsandbytes>=0.48.2", "accelerate>=1.2.1", "safetensors>=0.4.5", "huggingface_hub>=0.20.0", "pillow>=11.0.0", "numpy>=1.26.0", "requests>=2.32.0", "psutil>=5.9.0"]

[project.urls]
Repository = "https://github.com/EricRollei/Comfy_HunyuanImage3"
#  Used by Comfy Registry https://registry.comfy.org
Documentation = "https://github.com/EricRollei/Comfy_HunyuanImage3/wiki"
"Bug Tracker" = "https://github.com/EricRollei/Comfy_HunyuanImage3/issues"

[tool.comfy]
PublisherId = "ericrollei"
DisplayName = "Comfy_HunyuanImage3"
Icon = ""
includes = [] 
# "requires-comfyui" = ">=1.0.0"  # ComfyUI version compatibility

